Restructured as per PR comments
